+++ multiselector-library/src/main/java/com/bignerdranch/android/multiselector/ModalMultiSelectorCallback.java
@@ -1,12 +1,12 @@
-package com.bignerdranch.android.recyclerviewchoicemode;
+package com.bignerdranch.android.multiselector;
 
 import android.support.v7.view.ActionMode;
 import android.view.Menu;
 
 /**
  * Abstract superclass for creating action mode callbacks that automatically toggle
- * a {@link com.bignerdranch.android.recyclerviewchoicemode.MultiSelector}'s
- * {@link com.bignerdranch.android.recyclerviewchoicemode.MultiSelector#setSelectable(boolean)}
+ * a {@link com.bignerdranch.android.multiselector.MultiSelector}'s
+ * {@link com.bignerdranch.android.multiselector.MultiSelector#setSelectable(boolean)}
  * value in {@link android.support.v7.view.ActionMode.Callback#onPrepareActionMode(android.support.v7.view.ActionMode, android.view.Menu)}
  * and {@link android.support.v7.view.ActionMode.Callback#onDestroyActionMode(android.support.v7.view.ActionMode)}.
  */
@@ -38,7 +38,7 @@ public abstract class ModalMultiSelectorCallback implements ActionMode.Callback 
     }
 
     /**
-     * Attached {@link com.bignerdranch.android.recyclerviewchoicemode.MultiSelector} instance.
+     * Attached {@link com.bignerdranch.android.multiselector.MultiSelector} instance.
      * @return Attached selector.
      */
     public MultiSelector getMultiSelector() {
@@ -46,7 +46,7 @@ public abstract class ModalMultiSelectorCallback implements ActionMode.Callback 
     }
 
     /**
-     * Attach this instance to a new {@link com.bignerdranch.android.recyclerviewchoicemode.MultiSelector}.
+     * Attach this instance to a new {@link com.bignerdranch.android.multiselector.MultiSelector}.
      * @param multiSelector Selector to attach.
      */
     public void setMultiSelector(MultiSelector multiSelector) {
