+++ multiselector-library/src/main/java/com/bignerdranch/android/multiselector/SingleSelector.java
@@ -1,4 +1,4 @@
-package com.bignerdranch.android.recyclerviewchoicemode;
+package com.bignerdranch.android.multiselector;
 
 public class SingleSelector extends MultiSelector {
     @Override
